# openapi.yaml
openapi: 3.0.3
info:
  title: Task Management Platform API
  description: |
    REST API for the Task Management & Analytics Platform.
    All protected endpoints require a Bearer JWT obtained from `POST /api/auth/login`.
  version: 1.0.0
  contact:
    name: TaskFlow API
servers:
  - url: https://task-management-laravel-api.onrender.com/api
    description: Production
  - url: http://localhost:8000/api
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:       { type: integer }
        name:     { type: string }
        email:    { type: string, format: email }
        role:     { type: string, enum: [admin, manager, member] }
        is_active: { type: boolean }
    Task:
      type: object
      properties:
        id:          { type: integer }
        title:       { type: string }
        description: { type: string, nullable: true }
        status:      { type: string, enum: [pending, in_progress, completed, cancelled] }
        priority:    { type: string, enum: [low, medium, high] }
        assigned_to: { $ref: '#/components/schemas/User' }
        created_by:  { $ref: '#/components/schemas/User' }
        team_id:     { type: integer }
        due_date:    { type: string, format: date-time, nullable: true }
        created_at:  { type: string, format: date-time }
        updated_at:  { type: string, format: date-time }
...
    Pagination:
      type: object
      properties:
        current_page: { type: integer }
        last_page:    { type: integer }
        per_page:     { type: integer }
        total:        { type: integer }
...
security:
  - bearerAuth: []
paths:
...
  /teams/{teamId}/tasks:
    get:
      tags: [Tasks]
      summary: List tasks for a team
      parameters:
        - { name: teamId,      in: path,  required: true, schema: { type: integer } }
        - { name: status,      in: query, schema: { type: string } }
        - { name: priority,    in: query, schema: { type: string } }
        - { name: assigned_to, in: query, schema: { type: integer } }
        - { name: page,        in: query, schema: { type: integer } }
        - { name: per_page,    in: query, schema: { type: integer } }
      responses:
        '200':
          description: Paginated task list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Task' } }
                  meta: { $ref: '#/components/schemas/Pagination' }
    post:
      tags: [Tasks]
      summary: Create a task (admin/manager only)
      parameters:
        - { name: teamId, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:       { type: string }
                description: { type: string }
                priority:    { type: string, enum: [low, medium, high] }
                assigned_to: { type: integer }
                due_date:    { type: string, format: date-time }
      responses:
        '201': { description: Task created }
        '403': { description: Forbidden }
  /tasks/{taskId}:
    get:
      tags: [Tasks]
      summary: Get task details
      parameters:
        - { name: taskId, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Task detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Task' }
    patch:
      tags: [Tasks]
      summary: Update a task
      parameters:
        - { name: taskId, in: path, required: true, schema: { type: integer } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:       { type: string }
                description: { type: string }
                priority:    { type: string }
                assigned_to: { type: integer }
                due_date:    { type: string, format: date-time }
      responses:
        '200': { description: Updated }
        '403': { description: Forbidden }
    delete:
      tags: [Tasks]
      summary: Delete a task
      parameters:
        - { name: taskId, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: Deleted }
        '403': { description: Forbidden }
...